backend/requirements.txt
# ── Core Framework ───────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# ── LangChain / LangGraph ───────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0

# ── Tools ────────────────────────────────────────────────────
httpx>=0.27.0
tavily-python>=0.5.0

# ── Cache & Database ────────────────────────────────────────
redis>=5.0.0
aiosqlite>=0.20.0

# ── Utilities ────────────────────────────────────────────────
Pillow>=10.0.0
python-dotenv>=1.0.0

# ── Dev / Testing ────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.3.0
